ai_engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.0
python-dotenv==1.0.0
requests==2.31.0
